templates shared between kumactl and kuma-dp
